+++ src/main/kotlin/serie2/problem/ProcessPointsCollectionsI2.kt
@@ -13,8 +13,8 @@ object ProcessPointsCollectionsI2 {
         readerfile1.forEachLine { l ->
             val i = l.trim().split(" ")
             if ( "v" == i[0] && i.size >= 4) { // v - i[0] , id - i [1], x - i[2]  e y - i[3] i.size = 4
-                val x = i[2].toInt()
-                val y = i[3].toInt()
+                val x = i[2].toFloat()
+                val y = i[3].toFloat()
                 val point = Point(x, y)
                 val list = map.get(point)
                 if (list == null) {
@@ -30,8 +30,8 @@ object ProcessPointsCollectionsI2 {
         readerfile2.forEachLine { l ->
             val i = l.trim().split(" ")
             if ( "v" == i[0] && i.size >= 4) { // v - i[0] , id - i [1], x - i[2]  e y - i[3] i.size = 4
-                val x = i[2].toInt()
-                val y = i[3].toInt()
+                val x = i[2].toFloat()
+                val y = i[3].toFloat()
                 val point = Point(x, y)
                 val list = map.get(point)
                 if (list == null) {
